requirements.txt
numpy==1.24.4
torch==2.1.1
attrdict==2.0.1
diffusers==0.24.0
Flask==2.2.5
imageio==2.9.0
matplotlib==3.7.2
mesh_to_depth==0.1.2
open3d==0.16.0
openai==0.27.9
opencv_python==4.8.0.76
opencv_python_headless==4.5.5.64
Pillow==9.5.0
PyYAML==6.0.1
Requests==2.31.0
salesforce_lavis==1.0.2
scipy==1.9.1
termcolor==2.4.0
h5py>=2.10.0
git+https://github.com/aleju/imgaug
git+https://github.com/jamesbowman/openexrpython.git
oyaml>=0.9
pathlib>=1.0.1
scikit-image>=0.15.0
Shapely>=1.6.4.post2
tensorboardX>=1.9
tensorflow>=1.14.0
termcolor>=1.1.0
tqdm>=4.36.1
protobuf==3.20.3